Restructure example a bit
